Pin the version of vtk
Wheels of vtku3dexporter are tightly coupled to the special vtk wheel
that is used in the process. For instance, it needs all the header files
of VTK. To prevent weird errors with pip/pipenv it is best to make sure
that at least the version is right. Only problem that remains is that
this package won't work at all with the PyPi vtk wheel.